% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA_search_connected.R
\name{GA_search_connected}
\alias{GA_search_connected}
\title{Genetic algorithm for module identification on weighted network}
\usage{
GA_search_connected(lambda, node_score, edge_score, EdgeList, num_iter = 1000,
  muCh = 0.05, zToR = 10, minsize = 30)
}
\arguments{
\item{lambda}{weight parameter in the objective}

\item{node_score}{A N-length vector storing the nodes}

\item{edge_score}{A M-length vector storing the edges}

\item{EdgeList}{A two-columns matrix with each row a pair of weighted edges}

\item{num_iter}{Number of iterations in GA}

\item{muCh}{Mutation rate in GA}

\item{zToR}{Zero to one ratio}

\item{minsize}{The minimal size of module}
}
\value{
a list containing module size, best score, module as a list of nodes 
and a list of GA functions.
}
\description{
Using genetic algorithm (GA) to identify active modules on weighted network. 
The connectedness of resulted module is ensured by connected components 
finding of binary encoded set of genes. Rewritten from COSINE package.
}
\examples{
library(COSINE)
data(scaled_node_score)
data(scaled_edge_score)
data(PPI)
GA_result <- GA_search_connected(lambda=0.5,scaled_node_score,
scaled_edge_score,PPI,num_iter=100, muCh=0.05, zToR=10, minsize=50)
GA_result2<-GA_search_PPI(lambda=0.5,scaled_node_score,scaled_edge_score,PPI, 
num_iter=100, muCh=0.05, zToR=10, minsize=50)
## visualized by igraph
selected = c()
for (i in 1:dim(PPI)[1]){
    if (PPI[i,1] \%in\% GA_result$Subnet || PPI[i,2] \%in\% GA_result$Subnet)
        selected = c(selected,i)
}
library(igraph)
g <- graph.data.frame(PPI[selected,], directed=FALSE)
V(g)$color <- "gray"
V(g)$color[match(GA_result$Subnet,V(g)$name)] <- "red"
layout <- layout.reingold.tilford(g, circular=T)
png(filename = 'Connected_PPI.png', width=4096, height = 3652)
plot(g,layout=layout, vertex.size=5, vertex.label.cex=5,vertex.label.dist=0.5,edge.width=5)
dev.off()

}
\author{
Dong Li, \email{dxl466@cs.bham.ac.uk}
}
\references{
Ma, Haisu, et al. "COSINE: COndition-SpecIfic sub-NEtwork 
identification using a global optimization method." 
Bioinformatics 27.9 (2011): 1290-1298.

Dong Li et al. Memetic algorithm for finding active connected subnetworks 
in intracellular networks. 2016.
}
\keyword{GA,}
\keyword{identification}
\keyword{module}

