% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPINet.R
\name{PPIFromBioGRID}
\alias{PPIFromBioGRID}
\title{PPI network from BioGRID}
\usage{
PPIFromBioGRID(pvalues, GeneNames, BioGRIDfilename, savename)
}
\arguments{
\item{pvalues}{A N-dimensional vector of gene expression level, processed by limma}

\item{GeneNames}{A N-dimensional vector of gene symbols, pre-filtering by expression profiles}

\item{BioGRIDfilename}{The file downloaded from BioGRID, like 
BIOGRID-ORGANISM-Homo_sapiens-3.4.138.tab.txt}

\item{savename}{The file to save R data for filtered gene symbols and network edgelist}
}
\description{
Constructing protenin-protein interaction network from BioGRID database
}
\author{
Dong Li, \email{dxl466@cs.bham.ac.uk}
}
\keyword{BioGRID}
\keyword{PPINet}

