% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPINet.R
\name{PPIFromString}
\alias{PPIFromString}
\title{PPI network from STRING}
\usage{
PPIFromString(pvalues, GeneNames, STRINGfilename, STRINGTheta = 0, savename)
}
\arguments{
\item{pvalues}{A N-dimensional vector of gene expression level, processed by limma}

\item{GeneNames}{A N-dimensional vector of gene symbols, pre-filtering by expression profiles}

\item{STRINGfilename}{The file downloaded from STRING, like 4932.protein.links.v10.txt}

\item{STRINGTheta}{The threshold for PPI edges selections when using STRING}

\item{savename}{The file to save R data for filtered gene symbols and network edgelist}
}
\description{
Constructing protenin-protein interaction network from STRING database, when 
STRING contains gene names from gene expression platform, like Saccharomyces 
cerevisiae
}
\examples{

}
\author{
Dong Li, \email{dxl466@cs.bham.ac.uk}
}
\keyword{PPINet}
\keyword{STRING}

