% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPINet.R
\name{PPIFromStringEnsembl}
\alias{PPIFromStringEnsembl}
\title{PPI network from STRING and Ensembl}
\usage{
PPIFromStringEnsembl(pvalues, GeneNames, STRINGfilename, STRINGTheta = 0,
  Ensemblfilename, savename)
}
\arguments{
\item{pvalues}{A N-dimensional vector of gene expression level, processed by limma}

\item{GeneNames}{A N-dimensional vector of gene symbols, pre-filtering by expression profiles}

\item{STRINGfilename}{The file downloaded from STRING, like 9606.protein.links.v10.txt}

\item{STRINGTheta}{The threshold for PPI edges selections when using STRING}

\item{Ensemblfilename}{The file downloaded from ensembl (Customise your download), 
like mart_export.txt, first column is Associated Gene Name like MT-ND1, 
second column is Ensembl Protein ID like ENSP00000354687}

\item{savename}{The file to save filtered gene symbols and network edgelist}
}
\description{
Constructing protenin-protein interaction network from STRING database, when 
protein name is needed for mapping with gene name from Ensembl
}
\author{
Dong Li, \email{dxl466@cs.bham.ac.uk}
}
\keyword{Ensembl}
\keyword{PPINet}
\keyword{STRING}

