\name{PPINet-package}
\alias{PPINet-package}
\alias{PPINet}
\docType{package}
\title{
\packageTitle{PPINet}
}
\description{
\packageDescription{PPINet}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PPINet}
\packageIndices{PPINet}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{PPINet}

Maintainer: \packageMaintainer{PPINet}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
