% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPINet.R
\name{PPIplot}
\alias{PPIplot}
\title{PPI network visulization}
\usage{
PPIplot(net, STRINGTheta = 0, savename)
}
\arguments{
\item{net}{Edgelist from \code{\link{PPIFromString}}, 
\code{\link{PPIFromStringEnsembl}} and \code{\link{PPIFromBioGRID}}}

\item{STRINGTheta}{The threshold for PPI edges selections when using STRING}

\item{savename}{Name for saving PPI edge list as plain file and also saving 
nodes in largest connected component and figure as eps format by default}
}
\description{
Plot protenin-protein interaction network by igraph
}
\examples{

}
\author{
Dong Li, \email{dxl466@cs.bham.ac.uk}
}
\keyword{CONSTRUCTION}
\keyword{STRING}

