% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPINet.R
\name{PPIFromBioGRID}
\alias{PPIFromBioGRID}
\title{PPI network from BioGRID}
\usage{
PPIFromBioGRID(pvalues, GeneNames, BioGRIDfilename, savename, column1, column2)
}
\arguments{
\item{pvalues}{A N-dimensional vector of gene expression level, processed by limma}

\item{GeneNames}{A N-dimensional vector of gene symbols, pre-filtering by expression profiles}

\item{BioGRIDfilename}{The file downloaded from BioGRID, like 
BIOGRID-ORGANISM-Homo_sapiens-3.4.138.tab.txt}

\item{savename}{The file to save filtered gene symbols and network edgelist}

\item{column1}{Source column in BioGRID file, for Saccharomyces_cerevisiae 
column1=1 and for Homo_sapiens column1=3}

\item{column2}{Target column in BioGRID file, for Saccharomyces_cerevisiae 
column2=3 and for Homo_sapiens column3=4}
}
\description{
Constructing protenin-protein interaction network from BioGRID database
}
\author{
Dong Li, \email{dxl466@cs.bham.ac.uk}
}
\keyword{BioGRID}
\keyword{PPINet}

